Narrow node parameters with an integer meaning from i32/f64 to i64

+++ editor/src/messages/portfolio/document/graph_operation/graph_operation_message_handler.rs
@@ -287,7 +287,7 @@ impl MessageHandler<GraphOperationMessage, GraphOperationMessageContext<'_>> for
 				// Insert the main chain node (Blend or Morph) depending on whether a blend count is provided, referencing
 				// its control path input by the Blend template's named position or the Morph proto node's parameter symbol
 				let (path_input_connector, layer_alias, path_alias) = if let Some(count) = blend_count {
-					let blend_node_id = modify_inputs.insert_blend_data(layer, count as f64);
+					let blend_node_id = modify_inputs.insert_blend_data(layer, count as i64);
 					(InputConnector::node_at_index(blend_node_id, BLEND_PATH_INPUT_INDEX), "Blend", "Blend Path")
 				} else {
 					let morph_node_id = modify_inputs.insert_morph_data(layer);
